[Silver III] Title: 주유소, Time: 320 ms, Memory: 41008 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -30,7 +30,7 @@ public class Main {
                 cMin = cities[i];
             else if (cities[i] < cMin)
                 cMin = cities[i];
-            output += cMin * distances[i];
+            output += (long) cMin * distances[i];
         }
         System.out.println(output);
